+++ pypeit/pypeit.py
@@ -284,8 +284,9 @@ class PypeIt:
         This will not crash if not all of the standard set of files are not provided
 
         Args:
-            run (bool, optional): If False, only print the calib names and do
-            not actually run.  Only used with the pypeit_parse_calib_id script
+            run (bool, optional):
+                If False, only print the calib names and do not actually run.
+                Only used with the ``pypeit_parse_calib_id`` script.
 
         Returns:
             dict: A simple dict summarizing the calibration names
@@ -360,6 +361,9 @@ class PypeIt:
 
         # Write
         msgs.info('Writing calib file')
+        # TODO: Why are we writing this in addition to the *.calib file (see the
+        # __init__ function)?  I think this *.calib_ids file is actually less
+        # informative and likely wrong...
         calib_file = self.pypeit_file.replace('.pypeit', '.calib_ids')
         ltu.savejson(calib_file, calib_dict, overwrite=True, easy_to_read=True)