+++ pypeit/scripts/parse_calib_id.py
@@ -18,9 +18,6 @@ class ParseCalibID(scriptbase.ScriptBase):
                                     width=width, formatter=argparse.RawDescriptionHelpFormatter)
         parser.add_argument('pypeit_file', type=str,
                             help='PypeIt reduction file (must have .pypeit extension)')
-    #    parser.add_argument('-c', '--cpus', default=False, action='store_true',
-    #                         help='Number of CPUs for parallel processing')
-
         return parser
 
     @staticmethod