+++ CHANGES.rst
@@ -4,6 +4,7 @@
 
 - Gemini/GMOS mask design slurping and usage
 - New GMOS wavelength solution
+- Added NIRES tutorial doc
 - reid_arxiv templates for all MMTO Blue Channel gratings and for MMTO Binospec G600 and G1000
 - Various bug fixes and enhancements to mmt_bluechannel and mmt_binospec support